No vdrive ROM reading.
It's not a real drive to execute code. Returning the ROM content is fooling
autodetection in user programs which incorrectly apply a fastloader. The
problem started when I fixed the ROM loading some time ago and it started to
return ROM content instead of zeros. Now back to zeros for compatibility.